split files/refactor: AssetsStorage & EngineRouter
