+++ src/pyved_engine/EngineRouter.py
@@ -9,11 +9,11 @@ import os
 import time
 from io import StringIO
 from math import degrees as _degrees
-
 from pyved_engine.sublayer_implem import GameEngineSublayer  # Step 3: Inject the dependency
-from . import dep_linking, _hub
+from . import dep_linking
 from . import pe_vars
 from . import state_management
+from .AssetsStorage import AssetsStorage
 from .actors_pattern import Mediator
 from .compo import gfx
 from .compo import vscreen
@@ -33,6 +33,7 @@ class EngineRouter:
         self.debug_mode = False
         self.low_level_service = sublayer_compo
         self.ready_flag = False
+        self.storage = AssetsStorage()
         # immediate bind
         self._hub = {
             'SpriteGroup': self.low_level_service.sprite.Group,
@@ -44,136 +45,6 @@ class EngineRouter:
     @staticmethod
     def get_game_ctrl():
         return _existing_game_ctrl
-
-    def preload_assets(self, adhoc_dict: dict, prefix_asset_folder, prefix_sound_folder,
-                       debug_mode=False, webhack=None) -> None:
-        """
-        expected to find the (mandatory) key 'images',
-        also we may find the (optionnal) key 'sounds'
-        """
-        if debug_mode:
-            print('*' * 50)
-            print(f' CALL to preload assets [webhack value:{webhack}]')
-            print('*' * 50)
-            print()
-
-        for asset_desc in adhoc_dict['asset_list']:
-            if isinstance(asset_desc, str):  # either sprsheet or image
-                kk = asset_desc.split('.')
-                # print('>>>>>charge file:', kk[0], kk[1])
-                # print('prefix_asset_folder?', prefix_asset_folder)
-
-                if kk[1] == 'json':
-                    y = prefix_asset_folder
-                    if webhack:
-                        y = webhack + prefix_asset_folder
-                    adhocv = None
-                    if webhack:
-                        if prefix_asset_folder == './':
-                            adhocv = ''
-                        else:
-                            adhocv = prefix_asset_folder
-                    pe_vars.spritesheets[kk[0]] = gfx.JsonBasedSprSheet(
-                        kk[0], pathinfo=y, is_webhack=adhocv
-                    )
-
-                elif kk[1] == 'ttf':  # a >single< TTF font
-                    key = "custom_ft"
-                    ft_filename = asset_desc
-
-                    if webhack:
-                        y = webhack + ft_filename
-                    else:
-                        y = prefix_asset_folder + ft_filename
-                    print('fetching font:', key, ft_filename, f'[{y}]')
-                    pe_vars.fonts[key] = self.low_level_service.new_font_obj(
-                        y,
-                        pe_vars.DATA_FT_SIZE
-                    )
-
-                else:  # necessarily an image
-                    if prefix_asset_folder == './':
-                        filepath = asset_desc
-                    else:
-                        filepath = prefix_asset_folder + asset_desc
-                    print('fetching image:', kk[0], filepath)
-                    pe_vars.images[kk[0]] = self.low_level_service.image_load(filepath)
-
-        # -------------------------
-        # loading sfx files
-        # -------------------------
-        for snd_elt in adhoc_dict['sound_list']:
-            k = snd_elt.split('.')[0]
-            filepath = prefix_sound_folder + snd_elt
-            if webhack is not None:
-                filepath = webhack + filepath
-            print('fetching the sound:', k, filepath)
-            pe_vars.sounds[k] = dep_linking.pygame.mixer.Sound(filepath)
-
-        # -------------------------
-        # loading data files
-        # -------------------------
-        # For example:
-        # - cartridge/conversation.json, or
-        # -  my_map.ncsv
-
-        # TODO ! unification/debug.
-        #  Right now both assets & data_files can have a .TTF
-        prefix = 'cartridge/'
-
-        if webhack:
-            for dat_file in adhoc_dict['data_files']:
-                fp = os.path.join(webhack, dat_file)
-                k, ext = dat_file.split('.')
-                if ext == 'json':  # not spritesheets! TODO control "data_hint?"
-                    with open(fp, 'r') as fptr:
-                        pe_vars.data[k] = json.load(fptr)
-                elif ext == 'ttf':
-                    pe_vars.data[k] = self.low_level_service.new_font_obj(fp, pe_vars.DATA_FT_SIZE)
-
-                elif ext == 'ncsv':
-                    print('»ncsv webhack:', webhack)
-                    _preload_ncsv_file(k, prefix, webhack_info=webhack)
-                else:
-                    print(f'*Warning!* Skipping data_files entry "{k}" | For now, only .TTF and .JSON can be preloaded')
-            return  # end special case implies we end processing, right there
-
-        for dat_file in adhoc_dict['data_files']:
-            k, ext = dat_file.split('.')
-            try:
-                # fresh filepath
-                filepath = prefix + dat_file
-
-                if ext == 'json':
-                    with open(filepath, 'r') as fptr:
-                        pe_vars.data[k] = json.load(fptr)
-                elif ext == 'ttf':
-                    pe_vars.data[k] = self.low_level_service.new_font_obj(filepath, pe_vars.DATA_FT_SIZE)
-
-                elif ext == 'ncsv':
-                    _preload_ncsv_file(k, prefix)
-
-                else:
-                    print(f'*Warning!* Skipping data_files entry "{k}" | For now, only .TTF and .JSON can be preloaded')
-
-            except FileNotFoundError:  # TODO refactor to detect case A/B right at the launch script? -->Cleaner code
-                new_prefix = os.path.join(_hub.bundle_name, prefix)
-                print('we modd the prefix -->', new_prefix)
-
-                # try again
-                # fresh filepath
-                filepath = new_prefix + dat_file
-                if webhack is not None:
-                    filepath = webhack + filepath
-                # ---
-
-                if ext == 'json':
-                    with open(filepath, 'r') as fptr:
-                        pe_vars.data[k] = json.load(fptr)
-                elif ext == 'ttf':
-                    pe_vars.data[k] = self.low_level_service.new_font_obj(filepath, pe_vars.DATA_FT_SIZE)
-                elif ext == 'ncsv':
-                    _preload_ncsv_file(k, new_prefix)
 
     @staticmethod
     def get_version():
@@ -325,6 +196,17 @@ class EngineRouter:
     def close_game(self):
         dep_linking.pygame.quit()
 
+    def close_game(self):
+        pe_vars.gameover = False
+
+        self.low_level_service.mixer.quit()
+        self.low_level_service.quit()
+
+        self.storage.images.clear()
+        self.storage.csvdata.clear()
+        self.storage.sounds.clear()
+        self.storage.spritesheets.clear()
+
     def surface_create(self, size):
         return dep_linking.pygame.surface.Surface(size)
 
@@ -368,60 +250,39 @@ def get_gs_obj(k):
 
 
 # vars
-_engine_rdy = False
 _upscaling_var = None
 _scr_init_flag = False
 
-
-def _preload_ncsv_file(filename_no_ext, file_prefix, webhack_info=None):
-    print('## attempts to load NCSV; args are:', filename_no_ext, file_prefix, webhack_info, '###')
-
-    # filepath = prefix_asset_folder + asset_desc if prefix_asset_folder else asset_desc
-    csv_filename = filename_no_ext + '.' + 'ncsv'
-
-    if webhack_info:
-        y = webhack_info + csv_filename
-    else:
-        y = file_prefix + csv_filename
-    print('>>>tryin to find data file:', y)
-    with open(y, 'r') as file:
-        str_csv = file.read()
-        f = StringIO(str_csv)
-        map_data = list()
-        reader = csv.reader(f, delimiter=',')
-        for row in reader:
-            if len(row) > 0:
-                map_data.append(list(map(int, row)))
-        pe_vars.csvdata[filename_no_ext] = map_data
-
-
-# --- rest of functions ---
-def bootstrap_e(maxfps=None, wcaption=None, print_ver_info=True):
-    global _engine_rdy
-    pe_vars.mediator = Mediator()
-
-    if maxfps is None:
-        y = 60
-    else:
-        y = maxfps
-    pe_vars.max_fps = y
-    # in theory the Pyv backend_name can be hacked prior to a pyv.init() call
-    # Now, let's  build a primal backend
-    v = pe_vars.ENGINE_VERSION_STR
-    if print_ver_info:
-        print(f'Booting up pyved-engine {v}...')
-
-    from .foundation.pbackends import build_primalbackend
-
-    # SIDE-EFFECT: Building the backend also sets kengi_inj.pygame !
-    _pyv_backend = build_primalbackend(pe_vars.backend_name)
-    # if you dont call this line below, the modern event system wont work (program hanging)
-    events.EvManager.instance().a_event_source = _pyv_backend
-
-    dep_linking.pygame.init()
-    if wcaption:
-        dep_linking.pygame.display.set_caption(wcaption)
-    _engine_rdy = True
+# --------------------------------
+#  old version
+#  need to keep, or we are good ?
+# TODO
+# def bootstrap_e(maxfps=None, wcaption=None, print_ver_info=True):
+#     global _engine_rdy
+#     pe_vars.mediator = Mediator()
+#
+#     if maxfps is None:
+#         y = 60
+#     else:
+#         y = maxfps
+#     pe_vars.max_fps = y
+#     # in theory the Pyv backend_name can be hacked prior to a pyv.init() call
+#     # Now, let's  build a primal backend
+#     v = pe_vars.ENGINE_VERSION_STR
+#     if print_ver_info:
+#         print(f'Booting up pyved-engine {v}...')
+#
+#     from .foundation.pbackends import build_primalbackend
+#
+#     # SIDE-EFFECT: Building the backend also sets kengi_inj.pygame !
+#     _pyv_backend = build_primalbackend(pe_vars.backend_name)
+#     # if you dont call this line below, the modern event system wont work (program hanging)
+#     events.EvManager.instance().a_event_source = _pyv_backend
+#
+#     dep_linking.pygame.init()
+#     if wcaption:
+#         dep_linking.pygame.display.set_caption(wcaption)
+#     _engine_rdy = True
 
 
 # -------------------------------
@@ -487,16 +348,6 @@ def _screen_param(gfx_mode_code, screen_dim, cached_paintev) -> None:
 
 
 _existing_game_ctrl = None
-
-def close_game():
-    pe_vars.gameover = False
-    dep_linking.pygame.mixer.quit()
-    dep_linking.pygame.quit()
-
-    pe_vars.images.clear()
-    pe_vars.csvdata.clear()
-    pe_vars.sounds.clear()
-    pe_vars.spritesheets.clear()
