+++ src/pyved_engine/__init__.py
@@ -10,23 +10,14 @@
 | (github.com/wkta) - Contact thomas@katagames.io    |
 +----------------------------------------------------+
 """
-from . import _hub
+from . import pe_vars as defs
 from .EngineRouter import EngineRouter
 from .Singleton import Singleton
 from ._classes import *
 from .compo import gfx
 from .compo import vscreen  # deprecated
 from .compo.GameTpl import GameTpl  # legacy cls
-from .foundation import legacy_evs  # we just copy the event system of pygame
 from .foundation.events import Emitter, EvListener, EngineEvTypes
-from . import pe_vars as defs
-
-
-# TODO remove this when we can
-# ive kept it for retro-compatibility with projects that target pyv v23.6a1
-# such as demos/ecs_naif or the very early stage pyved ships-with-GUI editor
-# from .utils._ecs_pattern import entity, component, System, SystemManager, EntityManager
-# from . import evsys0
 
 
 _stored_kbackend = None
