+++ src/pyved_engine/compo/gfx.py
@@ -5,7 +5,6 @@ from collections import defaultdict
 
 from . import packed_capello_ft
 from .SpriteSheet import SpriteSheet as JsonBasedSprSheet
-from .. import _hub
 
 
 class BaseCfont:
