+++ src/pyved_engine/foundation/pbackends.py
@@ -1,6 +1,5 @@
 from .defs import EngineEvTypes, to_camelcase
 from .interfaces import BaseKenBackend
-from .. import _hub
 from .. import pe_vars
 from .. import dep_linking
